% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.create_enrich_dotplot}
\alias{.create_enrich_dotplot}
\title{Create enrichment dotplot (internal)}
\usage{
.create_enrich_dotplot(
  res,
  show.term.num,
  x,
  color.by,
  size.by,
  colors,
  size.range,
  title,
  legend.position,
  theme
)
}
\description{
Create enrichment dotplot (internal)
}
\keyword{internal}
