% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.filter_protein_coding}
\alias{.filter_protein_coding}
\title{Filter to protein-coding genes}
\usage{
.filter_protein_coding(gene_lists, organism, min.genes = 10)
}
\description{
Filter to protein-coding genes
}
\keyword{internal}
