% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\name{.gpdb_batch_query_metadata}
\alias{.gpdb_batch_query_metadata}
\title{Query Dataset Metadata in Batch}
\usage{
.gpdb_batch_query_metadata(dataset_ids)
}
\arguments{
\item{dataset_ids}{Character vector. Dataset IDs}
}
\value{
Data frame. Metadata for all datasets
}
\description{
Query Dataset Metadata in Batch
}
\keyword{internal}
