% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{.gpdb_close_progress}
\alias{.gpdb_close_progress}
\title{Close Progress Bar}
\usage{
.gpdb_close_progress(pb)
}
\arguments{
\item{pb}{Progress bar object}
}
\description{
Close Progress Bar
}
\keyword{internal}
