% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_cache.R
\name{.gpdb_convert_to_gene_name}
\alias{.gpdb_convert_to_gene_name}
\title{Convert gene_id to gene_name in expression data (optimized)}
\usage{
.gpdb_convert_to_gene_name(expr_data)
}
\description{
Convert gene_id to gene_name in expression data (optimized)
}
\keyword{internal}
