% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gpdb_generate_summary}
\alias{.gpdb_generate_summary}
\title{Generate text summary}
\usage{
.gpdb_generate_summary(gene, n_datasets, top_up, top_down, pathways)
}
\description{
Generate text summary
}
\keyword{internal}
