% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_cache.R
\name{.gpdb_get_gene_mapping}
\alias{.gpdb_get_gene_mapping}
\title{Get Gene Annotation Mapping (with cache)}
\usage{
.gpdb_get_gene_mapping()
}
\description{
Get Gene Annotation Mapping (with cache)
}
\keyword{internal}
