% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{.gpdb_get_palette}
\alias{.gpdb_get_palette}
\title{Get Color Palette}
\usage{
.gpdb_get_palette(palette_name = "spectral", n = NULL)
}
\arguments{
\item{palette_name}{Character. Name of palette}

\item{n}{Integer. Number of colors needed (will interpolate if needed)}
}
\value{
Character vector of colors
}
\description{
Get Color Palette
}
\keyword{internal}
