% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.gpdb_ora_enrich}
\alias{.gpdb_ora_enrich}
\title{Local Over-Representation Analysis using Fisher's exact test}
\usage{
.gpdb_ora_enrich(
  genes,
  geneset_df,
  p.cutoff = 0.05,
  q.cutoff = 0.05,
  background.genes = NULL
)
}
\description{
Returns list(All, Sig) with standard ORA column names.
}
\keyword{internal}
