% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{.gpdb_theme_default}
\alias{.gpdb_theme_default}
\title{Theme and Color System for GenePerturbR}
\usage{
.gpdb_theme_default(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
ggplot2 theme
}
\description{
Unified theme and color palette system for all visualizations
}
\keyword{Default}
\keyword{GenePerturbR}
\keyword{Theme}
\keyword{internal}
