% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{.gpdb_update_progress}
\alias{.gpdb_update_progress}
\title{Update Progress Bar}
\usage{
.gpdb_update_progress(pb, value)
}
\arguments{
\item{pb}{Progress bar object}

\item{value}{Current value}
}
\description{
Update Progress Bar
}
\keyword{internal}
