% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.parse_enrich_input}
\alias{.parse_enrich_input}
\title{Parse input for enrichment analysis}
\usage{
.parse_enrich_input(targets, split.by, top_up, top_down)
}
\description{
Parse input for enrichment analysis
}
\keyword{internal}
