% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{gpdb_get_info}
\alias{gpdb_get_info}
\title{Get Dataset Information}
\usage{
gpdb_get_info(dataset_id)
}
\arguments{
\item{dataset_id}{Character. Dataset ID}
}
\value{
List with metadata
}
\description{
Get Dataset Information
}
