% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{gpdb_list_datasets}
\alias{gpdb_list_datasets}
\title{List Datasets}
\usage{
gpdb_list_datasets(
  gene = NULL,
  cell_line = NULL,
  tissue = NULL,
  method = NULL,
  min_quality = NULL,
  quality_tier = NULL
)
}
\arguments{
\item{gene}{Character. Gene symbol}

\item{cell_line}{Character. Cell line name}

\item{tissue}{Character. Tissue type}

\item{method}{Character. Perturbation method}

\item{min_quality}{Numeric. Not implemented in current version}

\item{quality_tier}{Character. Not implemented in current version}
}
\value{
Data frame with dataset information
}
\description{
List Datasets
}
