% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{gpdb_load_metadata}
\alias{gpdb_load_metadata}
\title{Load Sample Metadata Only}
\usage{
gpdb_load_metadata(dataset_id)
}
\arguments{
\item{dataset_id}{Character. Dataset ID}
}
\value{
Data frame
}
\description{
Load sample-level metadata.
For loading expression + metadata together, use gpdb_load_data().
}
